Fix: 'A'+'P' regression when dropping by type
Selecting both 'P' (the just-picked-up items category) and 'A' (the
'auto-select relevant items' flag) when dropping items by type with 'D'
was automatically dropping every item in inventory instead of only items
that were marked as just having been picked up.  Basically, it was
causing 'A' to act like it did before 182e403 amended its functionality.

This commit is more-or-less identical to 933e6df, both in terms of the
problem and the fix, except that it applies to dropping items instead of
looting.